+++ Develappers.BillomatNet/CHANGELOG.md
@@ -32,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - implemented service for U operations on one invoice
 - implemented service for C operations on invoice mail
 - implemented service for CUD operations on one invoice item
-- implemented service for CR operations on one invoice comment
+- implemented service for CRD operations on one invoice comment
 - implemented service for R operations on invoice comments
 - implemented service for CRD operations on one invoice payment
 - implemented service for R operations on invoice payments